vetup/requirements.txt
elasticsearch==7.8.0
Flask==1.1.1
Flask-WTF==0.14.2
gunicorn==20.0.4
Werkzeug==0.16.0
python-dotenv==0.10.3


#WTForms==2.2.1
# appdirs==1.4.3
# attrs==19.3.0
# beautifulsoup4==4.8.2
# breadability==0.1.20
# certifi==2019.11.28
# chardet==3.0.4
# Click==7.0
# docopt==0.6.2
# idna==2.8
# importlib-metadata==1.3.0
# itsdangerous==1.1.0
# Jinja2==2.10.3
# joblib==0.14.1
# lexrank==0.1.0
# lxml==4.4.2
# MarkupSafe==1.1.1
# more-itertools==8.0.2
# nltk==3.4.5
# numpy==1.18.0
# packaging==19.2
# pandas==0.25.3
# path==13.1.0
# path.py==12.4.0
# pdfminer==20191125
# Pillow==6.2.1
# pluggy==0.13.1
# py==1.8.1
# pycountry==19.8.18
# pycryptodome==3.9.4
# pyocr==0.7.2
# pypandoc==1.4
# pyparsing==2.4.6
# pyrsistent==0.15.6
# pytest==5.3.2
# pytest-dotenv==0.4.0
# python-dateutil==2.8.1
# pytz==2019.3
# regex==2019.12.20
# requests==2.22.0
# scikit-learn==0.22.1
# scipy==1.4.1
# selenium==3.141.0
# six==1.13.0
# sklearn==0.0
# soupsieve==1.9.5
# sumy==0.8.1
# tabulate==0.8.6
# uritools==3.0.0
# urlextract==0.14.0
# urllib3==1.25.7
# Wand==0.5.8
# wcwidth==0.1.7
# zipp==0.6.0
